+++ src/rest/defaults.ts
@@ -1,13 +1,19 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
  */
 export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
 [
+    BodyParameterMiddleware,
+    PathParameterMiddleware,
+    QueryObjectParameterMiddleware,
+    QueryParameterMiddleware,
+    // HeaderMethodMiddleware,
+    HeaderParameterMiddleware,
     ProducesMiddleware,
     AdditionalDataMiddleware,
     IgnoredInterceptorsMiddleware,