+++ src/rest/middlewares/index.ts
@@ -5,3 +5,7 @@ export * from './ignoredInterceptors.middleware';
 export * from './additionalData.middleware';
 export * from './produces.middleware';
 export * from './bodyParameter.middleware';
+export * from './pathParameter.middleware';
+export * from './queryParameter.middleware';
+export * from './queryObjectParameter.middleware';
+export * from './headerParameter.middleware copy';